refactored save credential and delete message functions from the message detail component to the message detail view
