+++ app/views/UserViews/UserMessageDetailView.tsx
@@ -1,6 +1,7 @@
 import { VerifiableCredential } from '@veramo/core';
 import React from 'react';
-import { Button, ScrollView } from 'react-native';
+import { Alert, ScrollView } from 'react-native';
+import { Button } from 'react-native-elements';
 import { USER_CREDENTIAL_URL, USER_HANDLE_MESSAGE_TOKEN_URL } from '../../api.constants.';
 import { MessageDetail } from '../../components/messageDetail';
 import { IHandledMessage } from '../../interfaces/IMessageData.interface';
@@ -12,6 +13,36 @@ import { IHandledMessage } from '../../interfaces/IMessageData.interface';
  */
 export function UserMessageDetailView({navigation, route}: any){
 	const {item} = route.params;
+
+	// deletes a message from the database
+	const deleteMessage = async () => {
+		try {
+			// TODO: Add API call that actually deletes the message.
+			navigation.navigate('UserMessages');
+			return Alert.alert('Deleted', 'The message was deleted');
+		} catch (error) {
+			console.error(error);
+			return alert('unable to remove message, please try again later');
+		}
+	};
+
+	// displays an alert message which forces the user to confirm the deletion.
+	const deleteMessageAlert = () => {
+		Alert.alert(
+			'Delete', 
+			'Are you sure you want to delete this message?',
+			[
+				{
+					text: 'Cancel',
+					onPress: () => {return;}
+				},
+				{
+					text: 'Delete',
+					onPress: deleteMessage
+				},
+			]
+		);
+	};
 
 	// save credentials to db
 	const saveCredentials = async (credentials: VerifiableCredential[]) => {
@@ -44,6 +75,7 @@ export function UserMessageDetailView({navigation, route}: any){
 
 			if (!rawdata.ok) {
 				alert('no credentials found in the message');
+				return;
 			}
 
 			const data: IHandledMessage = await rawdata.json();
@@ -57,8 +89,9 @@ export function UserMessageDetailView({navigation, route}: any){
 
 	return (
 		<ScrollView>
-			<MessageDetail navigation={navigation} item={item}></MessageDetail>
-			<Button onPress={handleMessageToken} title='Save credential' color={'orange'}></Button>
+			<MessageDetail item={item}></MessageDetail>
+			<Button buttonStyle={{marginVertical: 5, marginHorizontal: 15, backgroundColor: 'orange'}} onPress={handleMessageToken} title='Save credential'></Button>
+			<Button buttonStyle={{marginVertical: 5, marginHorizontal: 15}} title='Delete' onPress={deleteMessageAlert}></Button>
 		</ScrollView>
 	);
 }
